Consolidate RL descriptors in metrics
Keep TensorBoard log conventions beside the parsing helpers and remove
the redundant rllib_descriptor module.
