+++ source/isaaclab/changelog.d/antoiner-benchmark-core.major.rst
@@ -3,8 +3,8 @@ Added
 
 * Added a backend-agnostic benchmark core under :mod:`isaaclab.test.benchmark`,
   including the ``capture``, ``metrics``, ``builders``, ``stepping``,
-  ``profiling``, and ``rllib_descriptor`` submodules, for assembling and
-  emitting schema-v1 benchmark bundles (``RuntimeBundle`` / ``TrainingBundle`` /
+  and ``profiling`` submodules, for assembling and emitting schema-v1 benchmark
+  bundles (``RuntimeBundle`` / ``TrainingBundle`` /
   ``StartupBundle``).
 * Added a ``schema`` output formatter that serializes a benchmark bundle through
   :class:`~isaaclab.test.benchmark.BaseIsaacLabBenchmark`, and taught
